services/heuristics/app.py
"""
Heuristics service for Smartrobe.

Extracts basic visual attributes using rule-based analysis and computer vision.
Currently implements color detection through image processing techniques.
"""

import asyncio
import random
import time
from typing import Any

from PIL import Image

from shared.base_service import CapabilityService
from shared.schemas import Color


class HeuristicsService(CapabilityService):
    """Heuristics service for rule-based attribute extraction."""

    SUPPORTED_ATTRIBUTES = {
        "color": Color,
    }
    SERVICE_TYPE = "heuristic"

    def __init__(self):
        super().__init__("heuristics", "heuristic", "1.0.0")

    async def _initialize_service(self) -> None:
        """Initialize heuristics service."""
        # In a real implementation, this would initialize:
        # - Color analysis models (K-means clustering, color space conversion)
        # - Image preprocessing pipelines
        # - Computer vision libraries (OpenCV, etc.)
        await asyncio.sleep(0.2)
        
        self.logger.info("Heuristics service initialized successfully")
        self.set_health_detail("color_detector_loaded", True)
        self.set_health_detail("image_processor_ready", True)
        self.set_health_detail("cv_libraries_loaded", True)

    async def _cleanup_service(self) -> None:
        """Cleanup heuristics service."""
        self.logger.info("Heuristics service cleaned up")

    async def extract_single_attribute(
        self, request_id: str, attribute_name: str, image_paths: list[str]
    ) -> tuple[Any, float]:
        """
        Extract a single attribute using heuristic methods.

        Args:
            request_id: Unique request identifier
            attribute_name: Name of the attribute to extract
            image_paths: List of paths to images

        Returns:
            Tuple of (attribute_value, confidence_score)
        """
        self.logger.info(
            f"Extracting {attribute_name} using heuristics",
            request_id=request_id,
            attribute=attribute_name,
            image_count=len(image_paths),
        )

        if attribute_name == "color":
            return await self._extract_color(request_id, image_paths)
        else:
            raise ValueError(f"Unsupported attribute: {attribute_name}")

    async def _extract_color(self, request_id: str, image_paths: list[str]) -> tuple[Color, float]:
        """
        Extract dominant color using heuristic color analysis.

        In a real implementation, this would:
        1. Load and preprocess images
        2. Convert to appropriate color space (HSV/LAB)
        3. Apply K-means clustering or histogram analysis
        4. Determine dominant color(s)
        5. Map to predefined color categories
        6. Calculate confidence based on color distribution

        Args:
            request_id: Unique request identifier
            image_paths: List of paths to images

        Returns:
            Tuple of (Color, confidence_score)
        """
        # Simulate processing time for color analysis
        processing_time = random.uniform(0.1, 0.3)
        await asyncio.sleep(processing_time)

        # Mock color detection with weighted probabilities
        # In reality, this would analyze actual image pixels
        color_probabilities = {
            Color.BLACK: 0.15,
            Color.WHITE: 0.12,
            Color.BLUE: 0.15,
            Color.RED: 0.1,
            Color.GRAY: 0.12,
            Color.NAVY: 0.08,
            Color.GREEN: 0.06,
            Color.BROWN: 0.08,
            Color.BEIGE: 0.05,
            Color.PINK: 0.04,
            Color.YELLOW: 0.03,
            Color.PURPLE: 0.02,
        }

        # Select color based on probabilities (mock implementation)
        colors = list(color_probabilities.keys())
        weights = list(color_probabilities.values())
        selected_color = random.choices(colors, weights=weights)[0]

        # Generate confidence score based on "color dominance"
        # Higher confidence for colors like black, white, blue (easier to detect)
        base_confidence = random.uniform(0.75, 0.95)
        if selected_color in [Color.BLACK, Color.WHITE, Color.BLUE, Color.RED]:
            confidence = min(0.98, base_confidence + random.uniform(0.05, 0.15))
        else:
            confidence = base_confidence

        self.logger.info(
            "Color extraction completed",
            request_id=request_id,
            detected_color=selected_color.value,
            confidence=confidence,
            processing_time_ms=int(processing_time * 1000),
        )

        return selected_color, confidence

    async def _check_service_health(self) -> bool:
        """Check heuristics service health."""
        try:
            # Test basic color detection capability
            test_start = time.time()
            await asyncio.sleep(0.01)  # Simulate quick health check
            
            health_check_time = time.time() - test_start
            self.set_health_detail("last_health_check_ms", int(health_check_time * 1000))
            self.set_health_detail("color_detection_ready", True)
            
            return True
        except Exception as e:
            self.logger.error("Heuristics service health check failed", error=str(e))
            return False


def create_app():
    """Create the heuristics FastAPI application."""
    service = HeuristicsService()
    return service.app


# Create app instance for uvicorn
app = create_app()


# For development/testing
if __name__ == "__main__":
    service = HeuristicsService()
    service.run()
