services/heuristics/pyproject.toml
[project]
name = "smartrobe-heuristics"
version = "0.1.0"
description = "Heuristic model service - rule-based attribute extraction from clothing images"
requires-python = ">=3.11,<3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "httpx",
    "pydantic",
    "pillow",
    "loguru",
    "python-multipart",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
include = ["*.py"]